+++ src/deberta_v2/deberta_v2_model.rs
@@ -118,8 +118,7 @@ impl FromStr for NormRelEmbedType {
         match s {
             "layer_norm" => Ok(NormRelEmbedType::layer_norm),
             _ => Err(RustBertError::InvalidConfigurationError(format!(
-                "Layer normalization type `{}` not in accepted variants (`layer_norm`)",
-                s
+                "Layer normalization type `{s}` not in accepted variants (`layer_norm`)",
             ))),
         }
     }
@@ -594,9 +593,12 @@ impl DebertaV2ForSequenceClassification {
     /// let device = Device::Cpu;
     /// let p = nn::VarStore::new(device);
     /// let config = DebertaV2Config::from_file(config_path);
-    /// let model = DebertaV2ForSequenceClassification::new(&p.root(), &config);
+    /// let model = DebertaV2ForSequenceClassification::new(&p.root(), &config).unwrap();
     /// ```
-    pub fn new<'p, P>(p: P, config: &DebertaV2Config) -> DebertaV2ForSequenceClassification
+    pub fn new<'p, P>(
+        p: P,
+        config: &DebertaV2Config,
+    ) -> Result<DebertaV2ForSequenceClassification, RustBertError>
     where
         P: Borrow<nn::Path<'p>>,
     {
@@ -613,7 +615,11 @@ impl DebertaV2ForSequenceClassification {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
 
         let classifier = nn::linear(
@@ -623,12 +629,12 @@ impl DebertaV2ForSequenceClassification {
             Default::default(),
         );
 
-        DebertaV2ForSequenceClassification {
+        Ok(DebertaV2ForSequenceClassification {
             deberta,
             pooler,
             classifier,
             dropout,
-        }
+        })
     }
 
     /// Forward pass through the model
@@ -660,7 +666,7 @@ impl DebertaV2ForSequenceClassification {
     /// # let device = Device::Cpu;
     /// # let vs = nn::VarStore::new(device);
     /// # let config = DebertaV2Config::from_file(config_path);
-    /// # let model = DebertaV2ForSequenceClassification::new(&vs.root(), &config);
+    /// # let model = DebertaV2ForSequenceClassification::new(&vs.root(), &config).unwrap();;
     /// let (batch_size, sequence_length) = (64, 128);
     /// let input_tensor = Tensor::rand(&[batch_size, sequence_length], (Kind::Int64, device));
     /// let mask = Tensor::zeros(&[batch_size, sequence_length], (Kind::Int64, device));
@@ -746,7 +752,10 @@ impl DebertaV2ForTokenClassification {
     /// let config = DebertaV2Config::from_file(config_path);
     /// let model = DebertaV2ForTokenClassification::new(&p.root(), &config);
     /// ```
-    pub fn new<'p, P>(p: P, config: &DebertaV2Config) -> DebertaV2ForTokenClassification
+    pub fn new<'p, P>(
+        p: P,
+        config: &DebertaV2Config,
+    ) -> Result<DebertaV2ForTokenClassification, RustBertError>
     where
         P: Borrow<nn::Path<'p>>,
     {
@@ -757,7 +766,11 @@ impl DebertaV2ForTokenClassification {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
         let classifier = nn::linear(
             p / "classifier",
@@ -766,11 +779,11 @@ impl DebertaV2ForTokenClassification {
             Default::default(),
         );
 
-        DebertaV2ForTokenClassification {
+        Ok(DebertaV2ForTokenClassification {
             deberta,
             dropout,
             classifier,
-        }
+        })
     }
 
     /// Forward pass through the model
@@ -802,7 +815,7 @@ impl DebertaV2ForTokenClassification {
     /// # let device = Device::Cpu;
     /// # let vs = nn::VarStore::new(device);
     /// # let config = DebertaV2Config::from_file(config_path);
-    /// # let model = DebertaV2ForTokenClassification::new(&vs.root(), &config);
+    /// # let model = DebertaV2ForTokenClassification::new(&vs.root(), &config).unwrap();
     /// let (batch_size, sequence_length) = (64, 128);
     /// let input_tensor = Tensor::rand(&[batch_size, sequence_length], (Kind::Int64, device));
     /// let mask = Tensor::zeros(&[batch_size, sequence_length], (Kind::Int64, device));