+++ src/mobilebert/mobilebert_model.rs
@@ -690,9 +690,12 @@ impl MobileBertForSequenceClassification {
     /// let device = Device::Cpu;
     /// let p = nn::VarStore::new(device);
     /// let config = MobileBertConfig::from_file(config_path);
-    /// let mobilebert = MobileBertForSequenceClassification::new(&p.root(), &config);
+    /// let mobilebert = MobileBertForSequenceClassification::new(&p.root(), &config).unwrap();
     /// ```
-    pub fn new<'p, P>(p: P, config: &MobileBertConfig) -> MobileBertForSequenceClassification
+    pub fn new<'p, P>(
+        p: P,
+        config: &MobileBertConfig,
+    ) -> Result<MobileBertForSequenceClassification, RustBertError>
     where
         P: Borrow<nn::Path<'p>>,
     {
@@ -703,7 +706,11 @@ impl MobileBertForSequenceClassification {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
         let classifier = nn::linear(
             p / "classifier",
@@ -711,11 +718,11 @@ impl MobileBertForSequenceClassification {
             num_labels,
             Default::default(),
         );
-        MobileBertForSequenceClassification {
+        Ok(MobileBertForSequenceClassification {
             mobilebert,
             dropout,
             classifier,
-        }
+        })
     }
 
     /// Forward pass through the model
@@ -748,7 +755,7 @@ impl MobileBertForSequenceClassification {
     /// # let device = Device::Cpu;
     /// # let vs = nn::VarStore::new(device);
     /// # let config = MobileBertConfig::from_file(config_path);
-    /// let model = MobileBertForSequenceClassification::new(&vs.root(), &config);
+    /// let model = MobileBertForSequenceClassification::new(&vs.root(), &config).unwrap();
     /// let (batch_size, sequence_length) = (64, 128);
     /// let input_tensor = Tensor::rand(&[batch_size, sequence_length], (Int64, device));
     /// let attention_mask = Tensor::zeros(&[batch_size, sequence_length], (Int64, device));
@@ -1127,9 +1134,12 @@ impl MobileBertForTokenClassification {
     /// let device = Device::Cpu;
     /// let p = nn::VarStore::new(device);
     /// let config = MobileBertConfig::from_file(config_path);
-    /// let mobilebert = MobileBertForTokenClassification::new(&p.root(), &config);
+    /// let mobilebert = MobileBertForTokenClassification::new(&p.root(), &config).unwrap();
     /// ```
-    pub fn new<'p, P>(p: P, config: &MobileBertConfig) -> MobileBertForTokenClassification
+    pub fn new<'p, P>(
+        p: P,
+        config: &MobileBertConfig,
+    ) -> Result<MobileBertForTokenClassification, RustBertError>
     where
         P: Borrow<nn::Path<'p>>,
     {
@@ -1140,7 +1150,11 @@ impl MobileBertForTokenClassification {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
         let classifier = nn::linear(
             p / "classifier",
@@ -1148,11 +1162,12 @@ impl MobileBertForTokenClassification {
             num_labels,
             Default::default(),
         );
-        MobileBertForTokenClassification {
+
+        Ok(MobileBertForTokenClassification {
             mobilebert,
             dropout,
             classifier,
-        }
+        })
     }
 
     /// Forward pass through the model
@@ -1185,7 +1200,7 @@ impl MobileBertForTokenClassification {
     /// # let device = Device::Cpu;
     /// # let vs = nn::VarStore::new(device);
     /// # let config = MobileBertConfig::from_file(config_path);
-    /// let model = MobileBertForTokenClassification::new(&vs.root(), &config);
+    /// let model = MobileBertForTokenClassification::new(&vs.root(), &config).unwrap();
     /// let (batch_size, sequence_length) = (64, 128);
     /// let input_tensor = Tensor::rand(&[batch_size, sequence_length], (Int64, device));
     /// let attention_mask = Tensor::zeros(&[batch_size, sequence_length], (Int64, device));