+++ src/prophetnet/prophetnet_model.rs
@@ -14,7 +14,7 @@ use std::borrow::Borrow;
 use std::collections::HashMap;
 
 use rust_tokenizers::tokenizer::{ProphetNetTokenizer, TruncationStrategy};
-use rust_tokenizers::vocab::{ProphetNetVocab, Vocab};
+use rust_tokenizers::vocab::ProphetNetVocab;
 use serde::{Deserialize, Serialize};
 use tch::{nn, Kind, Tensor};
 
@@ -1098,9 +1098,7 @@ impl
 
         let pad_token = match pad_token_id {
             Some(value) => value,
-            None => self
-                ._get_tokenizer()
-                .convert_tokens_to_ids(&[ProphetNetVocab::unknown_value()])[0],
+            None => self._get_tokenizer().get_unk_id(),
         };
 
         let token_ids = token_ids