+++ src/reformer/reformer_model.rs
@@ -548,7 +548,7 @@ impl ReformerModelWithLMHead {
         if let Some(lsh_num_chunks_after) = config.lsh_num_chunks_after {
             if config.attn_layers.contains(&AttentionType::lsh) & (lsh_num_chunks_after != 0) {
                 return Err(RustBertError::InvalidConfigurationError(
-                    format!("For text generation using LSH attention ensure `config.lsh_num_chunks_after` is set to 0 (currently {})", lsh_num_chunks_after),
+                    format!("For text generation using LSH attention ensure `config.lsh_num_chunks_after` is set to 0 (currently {lsh_num_chunks_after})"),
                 ));
             }
         }
@@ -556,7 +556,7 @@ impl ReformerModelWithLMHead {
         if let Some(local_num_chunks_after) = config.local_num_chunks_after {
             if config.attn_layers.contains(&AttentionType::local) & (local_num_chunks_after != 0) {
                 return Err(RustBertError::InvalidConfigurationError(
-                    format!("For text generation using local attention ensure `config.local_num_chunks_after` is set to 0 (currently {})", local_num_chunks_after),
+                    format!("For text generation using local attention ensure `config.local_num_chunks_after` is set to 0 (currently {local_num_chunks_after})"),
                 ));
             }
         }
@@ -709,14 +709,15 @@ impl ReformerClassificationHead {
             config.hidden_size,
             Default::default(),
         );
-        let num_labels = match &config.id2label {
-            Some(value) => value.len() as i64,
-            None => {
-                return Err(RustBertError::InvalidConfigurationError(
-                    "an id to label mapping must be provided for classification tasks".to_string(),
-                ));
-            }
-        };
+        let num_labels = config
+            .id2label
+            .as_ref()
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
+            .len() as i64;
         let out_proj = nn::linear(
             p / "out_proj",
             config.hidden_size,