+++ src/xlnet/xlnet_model.rs
@@ -919,7 +919,7 @@ impl XLNetForSequenceClassification {
     /// let device = Device::Cpu;
     /// let p = nn::VarStore::new(device);
     /// let config = XLNetConfig::from_file(config_path);
-    /// let xlnet_model = XLNetForSequenceClassification::new(&p.root(), &config);
+    /// let xlnet_model = XLNetForSequenceClassification::new(&p.root(), &config).unwrap();
     /// ```
     pub fn new<'p, P>(
         p: P,
@@ -936,7 +936,11 @@ impl XLNetForSequenceClassification {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
 
         let logits_proj = nn::linear(
@@ -1080,7 +1084,7 @@ impl XLNetForTokenClassification {
     /// let device = Device::Cpu;
     /// let p = nn::VarStore::new(device);
     /// let config = XLNetConfig::from_file(config_path);
-    /// let xlnet_model = XLNetForTokenClassification::new(&p.root(), &config);
+    /// let xlnet_model = XLNetForTokenClassification::new(&p.root(), &config).unwrap();
     /// ```
     pub fn new<'p, P>(
         p: P,
@@ -1095,7 +1099,11 @@ impl XLNetForTokenClassification {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
 
         let classifier = nn::linear(