+++ tests/deberta.rs
@@ -41,7 +41,7 @@ fn deberta_natural_language_inference() -> anyhow::Result<()> {
         false,
     )?;
     let config = DebertaConfig::from_file(config_path);
-    let model = DebertaForSequenceClassification::new(vs.root(), &config);
+    let model = DebertaForSequenceClassification::new(vs.root(), &config)?;
     vs.load(weights_path)?;
 
     //    Define input
@@ -170,7 +170,7 @@ fn deberta_for_token_classification() -> anyhow::Result<()> {
     dummy_label_mapping.insert(2, String::from("PER"));
     dummy_label_mapping.insert(3, String::from("ORG"));
     config.id2label = Some(dummy_label_mapping);
-    let model = DebertaForTokenClassification::new(vs.root(), &config);
+    let model = DebertaForTokenClassification::new(vs.root(), &config)?;
 
     //    Define input
     let inputs = ["Where's Paris?", "In Kentucky, United States"];